backend/requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.11
django-cors-headers==4.9.0
django-environ==0.13.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
magic-filter==1.0.12
multidict==6.7.1
packaging==26.0
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.23.0